Update card Flip
Change setFlipped value !flipped to true to ensure that the card is flipped back and prevent the card flipping back and forth when clicked multiple times.

Add flip timer so that when a card is flipped back, it is flipped front again in x seconds. This is to make the user pick a card again.