Fix card flip bug
Fix card flip bug that the front card momentarily shows when flipped. Before we flipped, the front becomes not visible and flipping the card is not as smooth, now we change that logic with backface-visibility: hidden; which is much smoother because theres no momentarily card invisibility.

Remove clickEvent on back card because it is automatically flipped to front.